pyproject.toml
[project]
name = "skills-demo"
version = "0.1.0"
description = "A web demo for writing, debugging and executing Claude Skills"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    # LLM Providers
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "google-generativeai>=0.8.0",
    "ollama>=0.3.0",
    # Storage
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    # Embeddings
    "voyageai>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
